+++ src/main/java/subway/persistence/repository/JdbcStationRepository.java
@@ -3,7 +3,6 @@ package subway.persistence.repository;
 import java.util.Optional;
 import org.springframework.stereotype.Repository;
 import subway.domain.Station;
-import subway.domain.StationRepository;
 import subway.persistence.dao.StationDao;
 import subway.persistence.entity.StationEntity;
