+++ src/main/java/subway/service/LineService.java
@@ -3,17 +3,17 @@ package subway.service;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import subway.domain.Line;
-import subway.domain.LineRepository;
 import subway.domain.Section;
 import subway.domain.Sections;
 import subway.domain.Station;
-import subway.domain.StationRepository;
 import subway.dto.request.AddStationToLineRequest;
 import subway.dto.request.DeleteStationFromLineRequest;
 import subway.dto.request.LineCreateRequest;
 import subway.exception.DuplicateLineException;
 import subway.exception.NotFoundLineException;
 import subway.exception.NotFoundStationException;
+import subway.persistence.repository.LineRepository;
+import subway.persistence.repository.StationRepository;
 
 @Service
 @Transactional
