+++ src/test/java/subway/persistence/repository/JdbcStationRepositoryTest.java
@@ -8,7 +8,6 @@ import org.junit.jupiter.api.DisplayNameGenerator.ReplaceUnderscores;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import subway.domain.Station;
-import subway.domain.StationRepository;
 import subway.persistence.RepositoryTest;
 
 @SuppressWarnings("NonAsciiCharacters")
